Created separate section for emissions in purchase contracts

+++ com-mmxlabs-lingo-reports/com.mmxlabs.lingo.reports/src/com/mmxlabs/lingo/reports/emissions/CargoEmissionAccountingReportJSONGenerator.java
@@ -76,9 +76,9 @@ public class CargoEmissionAccountingReportJSONGenerator{
 		model.baseFuelEmissionRate = vessel.getBaseFuelEmissionRate();
 		model.bogEmissionRate = vessel.getBogEmissionRate();
 		model.pilotLightEmissionRate = vessel.getPilotLightEmissionRate();
-		model.baseFuelEmission = 0.0;
-		model.bogEmission = 0.0;
-		model.pilotLightEmission = 0.0;
+		model.baseFuelEmission = 0;
+		model.bogEmission = 0;
+		model.pilotLightEmission = 0;
 		LocalDateTime eventStart = null;
 		
 		for (final Event e : cargoAllocation.getEvents()) {
@@ -100,13 +100,13 @@ public class CargoEmissionAccountingReportJSONGenerator{
 		for (final FuelQuantity fq : fuelQuantity) {
 			switch (fq.getFuel()) {
 			case BASE_FUEL: 
-				fq.getAmounts().forEach(fa -> model.baseFuelEmission += fa.getQuantity() * model.baseFuelEmissionRate);
+				fq.getAmounts().forEach(fa -> model.baseFuelEmission += (int) (fa.getQuantity() * model.baseFuelEmissionRate));
 				break;
 			case FBO, NBO:
-				fq.getAmounts().forEach(fa -> model.bogEmission += fa.getQuantity() * model.bogEmissionRate);
+				fq.getAmounts().forEach(fa -> model.bogEmission += (int) (fa.getQuantity() * model.bogEmissionRate));
 				break;
 			case PILOT_LIGHT:
-				fq.getAmounts().forEach(fa -> model.pilotLightEmission += fa.getQuantity() * model.pilotLightEmissionRate);
+				fq.getAmounts().forEach(fa -> model.pilotLightEmission += (int) (fa.getQuantity() * model.pilotLightEmissionRate));
 				break;
 			default:
 				throw new IllegalArgumentException("Unexpected value: " + fq.getFuel());
